Cmdline interface in Perl/fixed opt bug in C
Added a command line interface to the Perl script to allow specification of initial values, size of buffer and iterations for the benchmark module.
Fixed optimization bug of the C code, which auto-annihilation of the memory allocation and freeing of the buffer under O2 and higher optimization levels. Basically the compiler noticed that the buffer allocated was fred without being used and did away with both steps. A simple walk over the buffer with the loop fixed this optimization bug (yes compilers can act stupid by being smart)